Action Phv Constraints Solver (#5206)
This PR add a action constraint checker to validate PHV packing. When a candidate allocation list is proposed and passed to can_pack function in action phv constrait, a solver will be created to collect all move-based instructions for all actions related to fields in the candidate container. If both source and destination has been allocated, the solver will check and generate instruction based on the PHV allocation.