P4C-3762: Fixed handling of reinterpret casts of 0 ++ expression (#5197)
PHV allocation was failing when dealing with statements of the form (int<16>)(8w0 ++ bit8_0), i.e. a reinterpret cast of 0 ++  expression. It wasn't able to allocate the higher part with the 0 constant. BTW such statements are produced by midend from the statements like (int<a>)(bit<a>)(bit<b>), b < a.
The fix generates two set instructions, one for the 0 constant, and the other one for the expression on the right-hand side of the ++ operator.