P4C-3841: Added support of sliced operands in ternary condition (#5248)
A ternary condition in the form `A relop B ? A : B, where relop is one of the relational operators,
is handled in a special way by using a min/max instruction, which allows it to source both operands
from PHVs. This improvement adds support of sliced operands A and B. E.g. the following statement
is now supported:

hdr.f1[15:0] = (hdr.f1[15:0] <= hdr.f2[15:0]) ? hdr.f1[15:0] : hdr.f2[15:0];