Use custom ErrorReporter class to export errors/warnings to Event Logger (#5243)
This is integration of following p4c PR: p4lang/p4c#2804 and barefootnetworks/compiler-interfaces#127.

The purpose is to emit errors/warnings to both console and Event logger. That is achieved by subclassing ErrorReporter and putting that new BfErrorReporter inside BFNCompileContext (from which it is retrieved from inside ::error and ::warning functions).