P4C-3900: Internal compiler error and nothing else (#5361)
when there's a match table with an attached stateful table that can’t fit in the same stage. TablePlacement
splits it off into a separate ‘detached attached’ stage table that runs in a later stage than all of the
match stage tables.

Because the compiler do (always) allocate an enable bit for these ‘detached attached’ tables. It's OK to run
the detached attached table “unconditionally” in the same TableSeq as the match table. The enable bit will
ensure it does not actually do anything if the current match stage table doesn't match.

The problem being that when there are multiple such tables with detached attached tables, with complex control
nesting, table placement screws up the next-table control dependencies and chooses an invalid order for things.

The test P4 program of P4C-3900 hit this problem and had the "next table comes before" error.

With the help of the enable bit, the ‘detached attached’ stage table can work like the always_run_action table.

So what this PR doese is:
skip building any control/data dependency on these ‘detached attached’ stage table and remove them from the next
table candidate lists in DefaultNext and JbayNextTable.

P4C-3900 #assign zhengxing.li@intel.com

P4C-3900: Check the element exists in the map or not before getting it.