P4C-3925: Force phv allocation for ghost intrinsic metadata on non gh… (#5443)
Issue:

For a multipipe program having at least one pipe with and one without Ghost, ghost packets are not seen on the Ghost control.

Diagnosis:

Traditionally single pipe programs with ghost control are replicated on all 4 pipes in Tofino2. Hence this effectively enables Ghost on all 4 pipes. For a multipipe program with say 4 pipes and one with ghost (pipe0) and 3 without (pipe1, pipe2, pipe3), we get the scenario where ghost is enabled on only one pipe. Ghost packets work on a credit mechanism which works when TM sends ghost packets to all pipes.

Singe ghost is not enabled in pipe1/2/3 the tm_status_phv register is not set on these pipes. This register is disabled and the pipe mask is set to default value. This causes the credit mechanism to fail and no packets are seen on pipe0.

Solution:

The fix to overcome this limitation is to enable the tm_status_phv register and set the pipe mask to 0. More details can be found in the Jira.

Implementation:

In order to implement the above solution we need the following fixes in the compiler

Identify the scenario where we have a combination of ghost and non ghost pipes in a multipipe program
Details: Added options.ghost_pipes and options.pipes which are global and are set according to input program. A common function in common/utils.h - ghost_only_on_other_pipes(int pipe_id) can be used to check if current pipe needs to implement this fix. Note, each pipe is compiled separately.

Allocate phv for ghost intrinsic metadata on the non ghost pipes. This is required since the enabling tm_status_phv register also means we are setting the tm_status_phv.phv value. While there is no ghost control the pipe will still see ghost packets and parse the metadata before returning credits. This also implies the allocated PHV must be reserved only in the parser and should not be reused in ingress.
Details: Added an IR Node for Ghost Parser - IR::BFN::GhostParser which holds the intrinsic metadata to be allocated for Ghost. This is a fake parser since actual ghost parsing is implicit, however it helps PHV allocation mechanism to identify and allocate these metadata fields directly without making any exceptions.

Add Assembly syntax to convey mask information for tm_status_phv register
Details : Additional assembly syntax in parser ghost which now has phv and mask section with the necessary encodings to be written to the regs. The assembler is also compiled per pipe but in this case it does not need to have additional information about other pipes. It simply encodes the given info from bfa.

P4C-3925 #assign amresh.kulkarni@intel.com