P4C-4104: fix MergeInstructions (#5481)
When an action contains two instructions over same container that could be merged, but their source operands have conflicting alignment requirements, the compiler currently refuses to merge them which may lead to bfas to fail.

But the alignment requirements conflict should be resolved in the combined instruction because the source operands should be combined. Therefore, this change adds IMPOSSIBLE_ALIGNMENT action analysis error code to list of errors that should be ignored when deciding which instructions can be merged.