P4C-4123: Splitting the table according to the test requirement (#5547)
The stf test expect the dmac_dmac table to be split in two stages and each
stage should have 4 ways and each way should pack 2 entries per word. This
was the case previously but after some change in the compiler, this table
was no longer split resulting in a single stage with 4 ways and 2 entries
per word solution. Accessing the 8th part of the table was no longer
possible since this new layout only have parts 0..7.

This commit force the layout of the table to be similar to the original
one.

P4C-4123 #assign carl.geoffrion@intel.com