P4C-4120: support 8 pipelines in t3na architecture (#5515)
This commit brings support of compiling program with 8 pipelines for Tofino 3.

To achieve that, couple of changes in following components are needed:
* assembler
Currently compiler driver script (bf-p4c) uses assembler (bfas) in that way, that for each logical pipe in the program, the corresponding *.bfa file for that pipeline is processed by assembler as pipe 0 (pipe 0 is the default when --pipe command line argument is not used). Then we rely on the driver to adjust the addresses from the binary created by assembler, when it configures the pipelines in the device. Therefore we do not need to change the number of pipelines which assembler can process.
But assembler is also generating handles for parsers and tables, which contain encoded pipeline ID so that the handles are different for different pipelines. These handles currently have 4 bits dedicated for pipeline ID, so we do not need to change the format of handles. The pipeline ID for handles is passed to assembler using command line option `--table-handle-offset`. Maximum allowed value for this handle offset is changed to 16 to reflect maximum value which is allowed by space in the handles.
* architecture definition file (`bf-p4c/p4include/tofino3arch.p4`)
There were only 4 pipes before in the definition of Switch package, this is increased to 8 for Tofino 3.
Annotation `pkginfo` is fixed to have correct value for `arch` parameter and the version is initialized to 0.1.0. This annotation is not used anywhere at the moment, so this should not have any negative effect.
* schema in compiler_interfaces
File `schemas/power_schema.py` in compiler_interfaces repo contains a field `"pipe_number"` which supported up to 4 pipes. This is increased to 8 in PR https://github.com/intel-restricted/networking.switching.barefoot.compiler-interfaces/pull/134 . It also means that this PR depends on that PR (it has to be merged first, so that this PR could be merged).
* schema version number
As `power_schema.py` has changed, the version number of power schema is also updated. The code which uses logging class `Power_Schema_Logger` needs to provide schema version number when creating new logger object. This version number has to match the version number used in `power_schema.py`. Unfortunately the version number was hard-coded in the code which creates the logger object, so it had to be changed too. To avoid the need of such changes in the future a macro for power schema version is used in a similar way as it was already used for other schema versions (f.e. see resources schema version) using the definitions in `CMakeLists.txt` files.

New sample testing program for t3na architecture with 8 distinct P4 pipelines is added. Currently there is only a compile only test with this program, this test shall be extended to STF or PTF test.

Besides that, we also need to update scripts: `scripts/p4c-gen-conf` and `scripts/p4c-gen-bfrt-conf` so that the <program_name>.conf file which is generated by them is in correct format which is expected by the driver.
This change will be done in a separate commit together with new STF or PTF test for testing 8 pipelines on Tofino 3.

P4C-4120 #assign matej.vido@intel.com