P4C-4107: Removed InsertTableToResetInvalidatedHeaders from post-midend to avoid error in folded programs (#5498)
ExtractBridgeInfo creates a set of pipes for the purposes of bridged header analysis. The pipes are thrown away later (before reaching backend). ExtractBridgeInfo creates the set of pipes in the manner that it always assigns the gress that emits a bridged header as ingress and the gress that extracts it as egress. In the case of egress->ingress bridging, it effectively swaps ingress with egress. This is necessary for some subsequent passes that perform bridged header analysis. However, some of the subsequent passes are incompatible with this change. As a consequence, in the case of InsertTableToResetInvalidatedHeaders (sub-pass of ResetInvalidatedChecksumHeaders), it inserts an egress table (created to zero-init invalidated fields used in a checksum computation) into an ingress table sequence, which produces the error.

Since ResetInvalidatedChecksumHeaders is not necessary for the purposes of bridged header analysis, it is removed from post-midend. An STF test that covers the problematic case is added. The conditions that trigger the error are described in the comment at the beginning of the test P4 program.