P4C-4223: Shrink liverange of clot-allocated read-only fields (#5684)
If a field has been allocated to clot and (1) it is not modified in MAUs; (2) not digested; Then its physcical live range should ignore the last read in deparser, except for the clot adjustment case. See https://jira.devtools.intel.com/browse/P4C-4223 for more about this case.

This feature is enabled for table-first pass only.