P4C-4200: Added support of usage of more registers with the same width and addressing in a single action (#5672)
When more registers of different widths were used inside a single action, the compiler might crash with the Cannot unify type 'bit<M>' with type 'bit<N>' error.

In certain cases, the compiler now allows to use multiple registers in a single action. The registers have to be of the same width and they need to be addressed equally. If any of the two conditions is not met, a more descriptive error message is printed out.

In addition, when more than four actions try to access a register, a more descriptive error message is printed out. This is caused by a hardware limitation.