P4C-4231 and P4C-4237: Enable resource mode and backtracking and add temp vars allocation. (#5715)
This PR solves two JIRA.

P4C-4231: Enable resource mode and backtracking for table-first mode. We add an additional state in table summary, which is the ALT_TRY_ENHANCED_TP that tries to run placement with backtracking and resource mode enabled, using the last failed placement as reference.

P4C-4237: Because P4C-4231 has triggered many table splitting behaviors and table splitting was not supported in table-first mode. This PR also adds a new incremental phv allocation solely for table-first mode. It simply allocates PHV temp vars that are referenced in IR to containers that are empty in the specified live ranges, without calling the phv allocation.