P4C-4295: Merge invalid live ranges to avoid bugs in other passes. (#5748)
When we have a field that is written and never read (and the other way around) action analysis cannot correctly process the allocation because it will see multiple AllocSlices when querying writing to field X of table T. Fixing action analysis is complicated because it involves changing all places that query PhvInfo from expecting a single AllocSlice to expecting a set of AllocSlices that should have the same container slice with different live ranges.
Instead, we can just merge those invald live ranges into one to avoid the issue. The cons of this approach is that it may create longer than necessary live ranges, for write-only slices.