P4C-4296: add smart byte packing before trivial allocation for better ixbar usage (#5791)
Integrate TableFieldPackOptimization pass in PHV allocation. In alt-phv-alloc mode, before trivial allocation and normal PHV allocation, compiler will run TableFieldPackOptimization pass and get a list of PackChoice and every PackChoice is a list of fieldslice candidates that can possibly be packed into a slice list. And based on each PackChoice, compiler will try to merge some SuperClusters together so that fieldslice candidates in the PackChoice can be packed into a slice list, which is inserted in the merged SuperCluster. And a trivial allocation will be performed to check if this merged SuperCluster is legitimate. If yes, the change is commited. If not, all changes will be rolled back. At the end of the iteration, a new std::listPHV::SuperCluster* cluster_group is generated and it is ready for the actual trivial allocation or normal phv allocation.

P4C-4296 #assign changhao.wu@intel.com