P4C-1390: conversion of the sample_e2e primitive (#5885)
The sample_e2e primitive should be converted into the Tofino's
coalesced mirror feature. However, this feature has not been
properly implemented and tested, hence the conversion cannot
be actually done.

This revision changes the compiler to report more descriptive error
if the primitive is used by an old P4-14 program.

If the conversion is made in the V1 architecture the primitive is converted
to sample3 or sample4 primitives and these primitives cause
compilation error at the beginning of the midend - rejection of
the primitive is a target dependent decision, hence I considered
midend better place to do so.

If the conversion is made in the TNA architecture, it's rejected directly
in the conversion code - just for simplicity of the implementation.

There is one bonus in this revision: the bootstrap_bfn_compilers.sh script
now generates build/reconfigure script which invokes cmake with exactly
same arguments. It's useful if a CMakeList is changed and one doesn't
want to retype all bootstrap arguments again.