P4C-4469: extra constraint of PHV range of deparser parameters, and fix a auto-alias bug. (#5971)
Although there is 3-bit shift for deparser parameter, there is an extra constraint on the upper bound of container index.
If the metadata is 8 bits or less, then it must be allocated within bits [7:0] of any container.
If the metadata is 9-16 bits, then it must be allocated within bits [15:0] of a PHV16 pr PHV32.
if the metadata is 17-32 bits, then it can be allocated allocated within bits[31:0] a PHV32.

For example, Suppose there is a 3-bit metadata, hardware require all 3 bits must be allocated to bit[7:0] of a PHV (e.g. PHV16)
These are the valid allocation scheme: phv[2:0], phv[3:1], phv[4:2], phv[5:3], phv[6:4], phv[7:5], and these are not, phv[8:6], phv[9:7].

I combined this constraint with the shift constraint and add a validation pass in post-allocation validation.