P4C-2455: XOR based hash algorithms (#5911)
The extern Hash now accepts new hashing algorithms (enumeration values) XOR8, XOR16 and XOR32.
These algorithms take specified field list, they create a continuous block of data and then they cut this
block into pieces of specified width (8, 16 or 32 bits). These pieces are then XORed together. If the length
of the data block is not aligned with the width, appropriate number of zero bits is appended after the
end of the block.