P4C-3966: Disable split attached for TOFINO (#6007)
Splitting attached table entries is already disabled for TOFINO, however in this case the match part which sets the index for the indirect stateful was split in the first stage and actual execution in the next. This is not required as the table can be directly matched with stateful execution in the same (next) stage. 

The fix disables the match split which effectively moves all execution to the next stage.

Added an stf test which invokes scapy to create the packet as tested in jira.