P4C-4444: [NSA] Customer P4 for a hash solution fails with compiler bug error (#6009)
As the hashing engine takes inputs from the PHV crossbar, just fields associated with PHV
containers are allowed as inputs into the hasher extern - e.g. header members, structure
members, variables. There are some exceptions, e.g.:
* constants are allowed because they are computed into the hashing seeds,
*concatenation operator is handled exactly as a field list -> its operands are added into
the final field list.

This commit adds new midend pass NormalizeHashList, which detects disallowed expressions
in the hasher field list and replaces them by a temporary variable. The temporary variable
is declared in the beginning of appropriate control.

Initialization of the variables is placed right before statements in the apply block which
invoke the action the variable belongs to (right before a direct action invocation or
an application of a table which can contain the action). The initializations must be
placed in the apply block because the compiler cannot split one action into several
stages, so the initialization cannot be placed right before hasher's call.