P4C-4593: Fix selector1 stf test (#6110)
For the action selector stf test (selector1.stf), we have to specify the size of container for field b2, because our STF test code, for now, have hard-coded action_addr. This address needs to be changed if we allocate differently. For example, if we allocate b2 to 16-bit container, then it needs to be 0x200, for 8-bit, it's 0x100.