P4C-4490: Fix incorrect compilation of slice comparisons (#6101)
Changes address 2 issues:

- Slice inequality comparison compiling to equality comparison
At gateway.cpp:1150, shifting val resulted in zeroing it out and led to incorrect bitmasks in match. I think the instruction val <<= match_field.range().lo; was missing, based on the fact that the function BuildGatewayMatch::constant works similarly but includes that instruction.

- Allocation of slices used in multiple expressions
If a slice is used in multiple expressions, it might be allocated only once and cause incorrect compilation.
If a xor_with slice is allocated at gateway.cpp:834 and also needed in another expression, it will not be allocated again leading to incorrect compilation. I have added a check that should stop that from happening.

- Added p4c_4490.p4 PTF test.

- Added a unit test.