merging P4C-4397 branch 03 (PR #6142)
Primary feature: report # of ERRORs and # of WARNINGs [to std. error _only_, as voted by a majority out of 6 in a meeting today and later approved by Boris] just before the-compiler-proper exits.  This should be the last output from the-compiler-proper _except_ when there is a bug-check/crash/exception.

Secondary feature: get the file with the-compiler-proper`s "main" to also be ["cpplint"] linted.