P4C-4659: Faster slicing by propagate slicing decision at head and tail byte. (#6180)
This PR adds an optimization that
(1) In choosing the next slice list to split, we prefer slice lists that are closer to a byte that it's container size has been decided. It can potentially trim the search tree by materializing split decisions earlier and prune invalid slicing.
(2) After slicing, proactively split out the tail of slice list if the tail byte's container size has been decided. Similar as above, potentially a smaller search tree.
This optimization has fixed two profile of noviflow (p4c-1332) for table-first pass and p4c-2200 for normal pass. However, because it triggered some regressions, it is only enabled for alt-phv-alloc for now.