P4C-4478: Fix RegisterAction predicate returning >16 bits (#6203)
If predicate() returns >16 bits, the return value is now assigned to a slice of LSBs of the destination operand. This is done to make sure that sizes of both operands of the "set" instruction match up, which previously was not necessarily the case. MSBs of the destination operand are set to zero to make it consistent with a typical cast.

Added a compile_only test based on customer code to make sure it no longer causes an internal compiler error. Also added a PTF test of a simplified program to make sure the return value of predicate() is as expected.