P4C-4668: Fixed incorrect active_headers in pass (#6249)
* P4C-4668: Fixed incorrect active_headers in pass

In HeaderMutex pass, under certain situations, if a gateway (if
statement) with a hdr.isValid() was encountered, the active_headers
information would not drop the information that was only true in the
body of the if statement upon leaving the body.

This would lead to actions and tables immediately outside the if
statement body to have the wrong active_headers.

* P4C-4668: Adjusted failing tests

Bug fix in HeaderMutex restricts PHV allocation a bit. Even if it is a
very small nudge, this pushes 2 tests in our regression that were on the
limit and already impacted by the previous introduction of HeaderMutex
over the edge again. Adjusted their parameters to allow them to pass.