P4C-4465: Add checks for parser match attributes (#6279)
The assembler now reports an error if a profile is missing the
match_port or match_inband_metadata attributes, or if an analyzer_stage
rule is missing the match_w0, match_w1 or match_state (unless the
analyzer_stage is named) attributes. Since the compiler sets all of
these matches to "match all" by default, missing them in the ASM file is
most likely a bug.