P4C-4426: parser packing validator. (#6229)
This PR adds a new parser packing validator that catches all possible bit corruption issue due to parser extractions. However, we left some loopholes to be compatible with compiler's behavior on master. For example, if a field will be clobber because of extraction, we will skip it if the packing of the field with the extracted field is unavoidable.