Avoid using Boost_USE_STATIC_RUNTIME (#6342)
Setting this option to ON does not work on Ubuntu 20.04 with its boost from packages.
The documentation for this option says:

> Set to ON or OFF to specify whether to use libraries linked statically to the
> C++ runtime ("s" tag). Default is platform dependent.

This suggests that the difference is whether the boost (static) libraries get linked to C++ libraries statically or dynamically. I.e. whether the boost .a depends on libstdc++.so/.a (in the particular case of using GCC on Linux). Since we link the libstdc++ statically in our "static" builds, this does not really matter.

The option is actually shown to be set to OFF in cmake example for building with static Boost: https://cmake.org/cmake/help/latest/module/FindBoost.html#examples.
The boost libraries with static runtime are not available in Ubuntu 20.04. In Ubuntu 16.04, both versions are available. With the boost in p4factory, both versions are available too. Both configurations produce (together with -DBoost_USE_STATIC_LIBS=ON) a binary with statically-linked boost.