P4C-4799: tof1: set container valid for zero-initializations to hdr fields (#6372)
Set the valid bit on all header fields with zero initialization using the
validity register in PMArb. This applies to _all_ paths through the parser, so
may result in some validity bits being set when a zero-initialization is not
encountered.

Alternative approach (not taken) is to keep zero initializations to header
fields in Tofino 1. This would ensure that the container validity bit is set,
but it will result in multiple writes to a container for some programs.