P4C-4319: Extend bitwise instructions support (#6361)
Added support for bitwise instructions in actions that overwrite part of
a container, as long as the operands are a PHV source and a constant
source. 
For AND and XNOR instructions, bits in the constant source
corresponding to the overwritten fields are set to 1 to preserve the
fields' contents.
OR/XOR do not require this treatment because the constant source
is zero-extended by default.