P4C-4847: Adding validation for potential container conflict on trivial allocation (#6448)
The PHV Alternative Allocation used the same packing POV bit mechanism as
the original code which use previous table allocation pass to avoid
container conflict of valid bit. The problem being that in PHV Alternative
Allocation mode we don't have such thing when dealing with trivial
allocation which caused some program to no longer fit. The symptom of
that was tables that were not mutually exclusive were not able to be
placed in the same stage because of container conflict created by this POV
packing. To fix this issue, we reused the same packing validation as for
metadata field on trivial allocation to make sure POV bit definition have
at least 1 minstage of room to pack them into the same container.