P4C-4871: Fixing various issues on alternative PHV mode (#6495)
* P4C-4871: Fixing various issues on alternative PHV mode

Adding validation of reverse dependencies when trying to overlay a
container between two physical liverange. Originally we only looked at
the physical location which work fine when table allocation replay work
but can cause cross dependencies between table if that step fail. We now
look for conditional dependency if the overlay end and start at the same
stage and avoid cross dependency.

Fixing SRAMs and MAPRAMs compute in StageUseEstimate. Table with R/W
capabilities was not computed properly and the spare bank was not added
in the SRAMs required. Table with a single layout was also not credited
for their resources. This bug can lead to suboptimal compute for the
resource based allocation.

Fixing ATCAM entries handling in alternative PHV mode when doing a
replay. The original ATCAM entry size was was not captured properly which
caused the replay to create ATCAM of the wrong size. In most case the
replayed variant would be significantly smaller than the requested size.

Removing early SRAMs/MAPRAMs usage validation that is known to be
inaccurate sometime.

* P4C-4871: Fixing regression failure

* P4C-4871: Address PR comments