[P4C-4846] Action Analysis Fix - Allow deposit field instructions with background source (#6537)
For "set" instructions which can be a deposit field, the instruction can be reading from 2 sources where one of the source acts as a background source where it is also the destination container. This allows the container to be partially written by the other source.

Action analysis identifies this condition and does not flag the instruction as an ILLEGAL_OVERWRITE and fail validation.