P4C-4829: Report error on too many large constants (#6542)
Added a check to the compiler to make sure that the number of large constants
in register actions can fit on the device. Previously this was only checked
in the assembler.
Also moved the check for maximum number of register params to have both
of them in the same place.