P4C-4938: Add pragmas to p4c-4770.p4 test
These pragmas ensure that the egress::start parser state is not split
in alt-phv-alloc, to ensure the behavior matches normal phv
alloc. When the start state is split, it causes a parser match
register allocation failure, and these pragmas ensure it does not get
split by making sure the extractor limit for each container type is
not surpassed.