P4C-5044: Clear maps in init of MeterSetup (#6763)
Not clearing data in MeterSetup between applications could lead to an internal compiler error if the pre-color field of a meter is assigned a value from a @flexible header. In this case, the ingress/egress threads of a pipe can be swapped, which means that the pre-color field can appear different to the one saved in the update_pre_colors map when the pass is applied to another pipe.