P4C-3825: Improved LoweredParserState merging (#6535)
Now hoists child match operations to parent match if all children share
the same content, then merges lowered parser states as normal. Repeats
this operation until IR converges to optimal solution that uses the
least amount of transitions to exit parser.

Hoisting consists of moving all the contents (extracts, checksums,
counters, etc.) of children LoweredParserMatches into all their parent
LoweredParserMatches if and only if they respect the following
constraints, in that order:

1 All children must cover the full match space or be/contain an
  unconditional match. For example, if the parents' children are two
  matches match 0x**10 and match 0x**11, they are ineligible because
  there is an implicit reject transition if the value being matched is
  0x**0*. However, if the parents' children were match 0x**10, match
  0x**11, and match 0x**0*, then the full match space is covered and
  the children are eligible for this optimization.
2 All children must have exactly the same contents.
3 All their parents have enough space in their input buffers to accept
  the contents.
4 Adding the contents of the children into the parents cannot violate
  hardware constraints. Those constraints are (in no particular order):

- Common (all chips)
  - Cannot use more extractors than there are available
  - Cannot have more than one counter operation
  - Cannot save to the same match registers more than once
  - Cannot use the same checksum unit more than once
  - Cannot write to the same container more than once
- TF2/TF3
  - Cannot extract more than 2 constants
  - Cannot contain more than 2 CLOTs
  - If the parent already contains an extract to a specific CLOT (e.g.
    CLOT 1), cannot add another spilled CLOT extract to the same CLOT

Once the content of the children are moved, the empty children remain
so that the flow of the parser graph remains the same. Those empty
children can then often be merged with their parent by the
MergeLoweredParserStates pass, reducing the amount of transitions
required to reach the end of the parser. This cycle can be repeated
until the parser graph can no longer be reduced in size.