P4C-5090: fix extraction population order to fix n2w alignment (#6782)
* P4C-5090: fix extraction population order to fix n2w alignment

Tofino 1: populate the extractors from extrctor 0 for all extractor
sizes, except when populating the dummy checksum extract (see
MODEL-210). This fixes an issue causing narrow-to-wide alignment
failures in the case when we have 1 x 16/32b checksum validation result
and 1 x same-sized constant extract. In this scenario, the allocation
would previously put the constant in extractor 1, resulting in 3 values
being written to the FIFO instead of 2 (csum result, empty, constant).

The population order likely also caused some performance failures for
cases where the parser is on the borderline of parsing fast enough.

* P4C-5090: fix swizzling logic in n2w extracts

* P4C-5090: filter the pad source size

Prefer a from_idx during padding that points to a container that is same
size or smaller than the extractor. This avoids problems where the pad
points to a large container but doesn't have enough extractors to fill.

* P4C-5090: add extra comment about swizzling