P4C-5104: Fix issue if 2+ gateway rows lead to same tag (#6951)
If multiple gateway rows lead to the same tag, their expressions are
essentially OR'd together. In those cases, you cannot deterministically
conclude which expression matched, so if there are reads on header POV
bits in those expressions, you have to conservatively set them to be
UNKNOWN rather than stating that they are either surely ACTIVE or
INACTIVE.

Previously, only the last gateway row leading to a tag would be
evaluated, leading to HeaderMutex containing the wrong info when
visiting those tags. Fixed this issue