P4C-2886: handle header stacks in clots (#7044)
* P4C-2886: handle header stacks in clots

For non-unrolled parser loops, handle header stacks by incrementing the
clot tag each iteration through the loop.

Includes handling of:
 - stride values > 1
 - loops with multiple paths

When we have loops with multiple paths, ensure that we only put things
in CLOTs when all elements of the stack are CLOTable.

Also fixed an issue with strided allocation on JBAY: the parser
containers don't increment consecutively by container type; instead
regular and mocha containers are interleaved, and for 8b containers, we
need to increment by 2 containers (since 2x8b containers fit in a single
16b parser container).

When a parser state can write to multiple stack indices due to different
paths through the parse graph, ensure that all or none of the indices
for that state are clottable.

Only adjust the container by the stack offset when the source is packet
data. If it is a constant, then the constant is rotated by the stack
offset instead of the container number being changed.