P4C-3892: delay parser shifts to allow match register alloc (#7110)
Identify defs used by select statements that potentially be delayed
until the following state. Delay the shift to allow the match register
to be allocated in the following state to improve the chances of a
successful alloc.

Report an error at the end of the final iteration if allocation was
unsuccessful, even if additional defs have been identified for delaying.

Don't attempt to delay the same def + destination in multiple rounds:
prevent the def from being delayed if it has been delayed previously.