initialize outputDir for gtest (#7153)
Initialize outputDir in the BFN_Options constructor do eliminate some
failures when running with high log levels in gtest. For example, this
cuases an exception to be thrown and the test to fail:

```
p4c/test/gtestp4c --gtest_filter=ClotTest.* -T allocate_clot:4
```

The reason is that at log level 3 or higher, AllocateClot invokes
DumpParsser, which uses the outputDir when constructing a path for the
parser graphs. If outputDir hasn't been initialized then an exception is
thrown with the message "basic_string::_M_construct null not valid".

[no-jira]