P4C-5102: varbit fix: mutliple start offsets (#7159)
* P4C-5102: varbit fix: mutliple start offsets

Support multiple different start/branch base offsets for a single varbit
field. This is necessary when we have code structured like the
following:

    state parse_ethernet {
        pkt.extract(hdr.ethernet);
        transition select(hdr.ethernet.src_addr[0:0]) {
            0 : parse_vb_even;
            1 : parse_vb_odd;
        }
    }

    state parse_vb_even {
        pkt.extract(hdr.vb, ((bit<32>)hdr.ethernet.dst_addr[3:1]) * 16);
        transition accept;
    }

    state parse_vb_odd {
        pkt.extract(hdr.vb, ((bit<32>)hdr.ethernet.dst_addr[3:1]) * 16 + 8);
        transition accept;
    }

Here we have two different states that extract into the same varbit
field. In the first state (parse_vb_even), we extract 16b chunks at
offsets of 0, 16, 32, etc., while in the second state (parse_vb_odd), we
extract 16b chunks at offsets of 8, 24, 40, etc.

The `create_branch_state` function walks over an ordered map of the
extracts until the header length exceeds the length of the varbit
extraction being processed. The problem is that in the scenario above,
the varbit_hdr_instance_to_header_types would end up with:
   [0, 16, 32, ..., 8, 24, 40, ...]
When processing the "odd" state (8, 24, 40, ... lengths), it would
incorrectly pick up the header types for the even state.

This change introduces an extra level of hierarcy (the base/start
offset) to the varbit_hdr_instance_to_header_types map to ensure that
the correct set of lengths is used.

* P4C-5102: address PR feedback from Chris