P4C-5288: improve csum swap update when shifting (#7176)
When shifting PacketRVals, improve the logic that identifies when a
checksum swap value needs to be inverted.

The original code identified the checksum to toggle when processing an
rval. The problem with this is that multiple checksums may share the
same rval, causing only the first checksum to be udpated. (The rval is
visited only once. We could enable multiple visits per node, but then
this would cause shared rvals to be uniquified when shifting.)

New code post-processes the checksum adds/subtracts and checks if the
rval alignment has changed, in which case the swap field is toggled.