P4C-5301: if simplification: avoid BlockStatement duplication (#7192)
Nested if simplification is intended to take a multi-level if statement
and convert it to a set of single-level if statements. For example, the
following nested if:

  if (condA)
      if (condB)
          // First block
      else
          // Second block

Should be simplified to a pair of single-level if statments:

  if (condA && condB)
      // First block
  if (condA && !condB)
      // Second block

The original code would create the two single-level if statements, but
it would also keep the original if statement. This occurred because the
algorithm recursively processed the substatement blocks and recorded the
combination of conditions required to execute each substatement block,
but it did not remove IfStatement objects within each substatement block
being recorded.

The fix removes IfStatement objects from any BlockStatement before
recording it with its corresponding set of conditions.